backen2/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
async_upnp_client==0.44.0
attrs==25.3.0
av==14.4.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==45.0.2
defusedxml==0.7.1
dnspython==2.7.0
fastapi==0.115.12
firebase-admin==6.8.0
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
ifaddr==0.2.0
msgpack==1.1.0
multidict==6.4.3
paho-mqtt==2.1.0
prometheus_client==0.22.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pyee==13.0.0
PyJWT==2.10.1
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-didl-lite==1.4.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
semver==3.0.4
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
voluptuous==0.15.2
watchdog==6.0.0
yarl==1.20.0
httpx

